Removed unnecessary asynchronous HTML loading, cleaned up code
Honesty not sure what I was thinking here when I decided it would be a good idea to have a function that downloads more HTML when I could just put it all in the same file and hide it? Way less race conditions and other DOM bullshit to deal with this way.